+++ rsearch/processor_test.go
